refactor: contributor type wrapper unified for contributor card and popover
