+++ src/components/v5/shared/Avatar/index.ts
@@ -1,2 +1,1 @@
 export { default } from './Avatar.tsx';
-export { default as AvatarWithStatusBadge } from './AvatarWithStatusBadge.tsx';
